refactor: move home-page schema to a new directory
