+++ schemas/home-page.ts
