+++ schemas/index.ts
@@ -1,19 +1,19 @@
 // TODO: Migrate schemas from './studio/schemas' to './schemas'
+import homePage from "schemas/home-page";
 import about from "studio/schemas/about";
 import accessibilityDeclaration from "studio/schemas/accessibility-declaration";
 import block from "studio/schemas/block";
 import characteristics from "studio/schemas/characteristics";
 import characteristics_list from "studio/schemas/characteristics_list";
+import collaborate from "studio/schemas/collaborate";
 import document from "studio/schemas/document";
 import documents from "studio/schemas/documents";
 import faq from "studio/schemas/faq";
-import homePage from "studio/schemas/home-page";
 import image from "studio/schemas/image";
 import link from "studio/schemas/link";
 import member from "studio/schemas/member";
 import toy from "studio/schemas/toy";
 import video from "studio/schemas/video";
-import collaborate from "studio/schemas/collaborate";
 
 const schemas = [
   /* Your types here! */
